strip/add prefixes
